\name{MyrrixClientConfiguration-class}
\alias{MyrrixClientConfiguration-class}
\title{Object of class MyrrixClientConfiguration}
\description{
  An object of class MyrrixClientConfiguration describes
  how access options to the ClientRecommender.
}
\section{Slots}{
  \describe{ \item{\code{config}:}{A java object
  representing a MyrrixClientConfiguration} }
}
\examples{
myconfig <- new("MyrrixClientConfiguration")
myconfig
}

